refactor: remove comments and change name
