+++ package.json
@@ -18,7 +18,6 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.3",
-    "axios": "^1.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.8",
