+++ src/components/Cards/Cards.tsx
@@ -1,6 +1,6 @@
 import { useAppDispatch, useTypedSelector } from '../../hooks/hooks';
 import { useGetBreedsQuery } from '../../store/api';
-import { setId, setVisibleModal } from '../../store/idSlice';
+import { setId, openModal } from '../../store/cardSlice';
 import { RootState } from '../../store/store';
 import { ICat } from '../../types/api.interface';
 import { Card } from '../card/Card';
@@ -9,12 +9,12 @@ import styles from './cards.module.css';
 
 export const Cards = () => {
   const dispatch = useAppDispatch();
-  const { value } = useTypedSelector((state: RootState) => state.filter);
+  const { value } = useTypedSelector((state: RootState) => state.cardSlice);
 
   const { isLoading, data } = useGetBreedsQuery(value);
 
   const handleClick = (id: string) => {
-    dispatch(setVisibleModal(true));
+    dispatch(openModal(true));
     dispatch(setId(id));
   };
