+++ src/components/catModal/catmodal.module.css
@@ -1,13 +1,3 @@
-/* .image > img {
-  display: block;
-  width: 100%;
-  height: 100%;
-  object-position: top;
-  object-fit: cover;
-  aspect-ratio: 1.5/1;
-  border-radius: 8px 8px 0px 0px;
-} */
-
 .overlay {
   position: fixed;
   width: 100%;
